# Use cached DNS_NAME for performance