--- a/CHANGES.rst
+++ b/CHANGES.rst
@@ -7,7 +7,7 @@ Change log
 * Added ``MAILER_USE_FILE_LOCK`` setting to allow disabling file based locking.
 * Added ``-r`` option to ``purge_mail_log`` management command. Thanks julienc91
 * Fixed deprecation warnings on Django 3.1
-
+* Use cached DNS_NAME for performance
 
 2.0.1 - 2020-03-01
 ------------------